# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: SuperNØva Smart Locker API
  description: |
    API RESTful per il sistema di smart locker SuperNØva.

    ## Autenticazione
    L'API utilizza JWT (JSON Web Tokens) per l'autenticazione. Includi il token nell'header:
    ```
    Authorization: Bearer <token>
    ```

    ## Rate Limiting
    - 100 richieste per 15 minuti per IP
    - Risposta 429 quando il limite viene superato

    ## Errori
    Tutte le risposte di errore seguono il formato:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Descrizione dell'errore"
      }
    }
    ```
  version: 1.0.0
  contact:
    name: SuperNØva Support
    email: api@supernova-locker.it
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.supernova-locker.it/v1
    description: Production server
...
paths:
...
  /locations/{deviceId}:
    get:
      tags: [Locations]
      summary: Posizione corrente dispositivo
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/deviceId'
      responses:
        '200':
          description: Posizione corrente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
    post:
      tags: [Locations]
      summary: Aggiorna posizione (uso interno)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/deviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [latitude, longitude]
              properties:
                latitude:
                  type: number
                  example: 45.052
                longitude:
                  type: number
                  example: 9.695
                accuracy_m:
                  type: number
                  example: 2.5
      responses:
        '200':
          description: Posizione aggiornata
  /locations/{deviceId}/history:
    get:
      tags: [Locations]
      summary: Cronologia posizioni
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/deviceId'
        - name: days
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 90
            default: 7
      responses:
        '200':
          description: Cronologia posizioni
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Location'
  /locations/{deviceId}/geofence:
    post:
      tags: [Locations]
      summary: Configura geofence
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/deviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Geofence'
      responses:
        '200':
          description: Geofence configurato
    delete:
      tags: [Locations]
      summary: Rimuovi geofence
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/deviceId'
      responses:
        '200':
          description: Geofence rimosso
  /locations/nearby/search:
    get:
      tags: [Locations]
      summary: Cerca dispositivi vicini
      security:
        - bearerAuth: []
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: lng
          in: query
          required: true
          schema:
            type: number
        - name: radius_m
          in: query
          schema:
            type: number
            default: 500
      responses:
        '200':
          description: Dispositivi trovati

  # ============================================
  # Access Control Endpoints
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    deviceId:
      name: deviceId
      in: path
      required: true
      description: ID univoco del dispositivo
      schema:
        type: string
        example: dev_cassapanca_001
  schemas:
...
    Location:
      type: object
      properties:
        latitude:
          type: number
          example: 45.052
        longitude:
          type: number
          example: 9.695
        accuracy_m:
          type: number
          example: 2.5
        timestamp:
          type: string
          format: date-time
    Geofence:
      type: object
      properties:
        center:
          type: object
          properties:
            lat:
              type: number
            lng:
              type: number
        radius_m:
          type: number
          minimum: 10
          maximum: 10000
        alert_on_exit:
          type: boolean
          default: true
        alert_on_enter:
          type: boolean
          default: false
...